Move view episode event to page load
